book_management_system/requirements.txt
asgiref==3.8.1
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
setuptools==68.2.2
sqlparse==0.5.0
wheel==0.41.2
